UniversalFakeDetectV2/requirements.txt
matplotlib==3.10.0
numpy==1.26.4
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.45.2
